% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vicite.R
\name{setup_vicite}
\alias{setup_vicite}
\title{produce a SingleCellExperiment with totalVI quantifications of 10k PBMC data (excluding
5k PBMC data) with cells held in common with the OSCA ch12 CITE-seq example}
\usage{
setup_vicite()
}
\description{
produce a SingleCellExperiment with totalVI quantifications of 10k PBMC data (excluding
5k PBMC data) with cells held in common with the OSCA ch12 CITE-seq example
}
\examples{
ss = setup_vicite()
ss
}
